feat(account): rewire Account::{new,read,link,set_number} to blob storage
Phases 2–4 of the blob-credentials refactor (issue #51).

Adds two new modules:

- `src/account/migration.rs`: pure-rust legacy-key reader trait
  (`LegacyKeyReader`) and `migrate_from_legacy` that converts the old
  per-key shape into a populated `AccountCredentials`. Defensive
  parsing mirrors `Account::read`'s post-#50 behavior — unparseable
  numerics default; unknown enum values default; bad hosts default
  to "signal.org". 22 unit tests including the iter-A.2.2
  hardware-observed contamination shape.

- `src/account/storage.rs`: `CredentialsStore` trait abstracting blob
  + legacy + sync, `PddbCredentialsStore` adapter wrapping a Pddb
  reference, `read_or_migrate` orchestration (blob → legacy fallback
  → fresh), `persist_credentials` for the post-link write. 12 unit
  tests using a `RefCell`-backed mock (fault injection: failed
  blob reads, failed writes; idempotence on re-run; corrupt-blob
  fallback). Tests fully cover the orchestration without spinning
  up real PDDB.

Rewrites `src/account.rs`:

- `Account::new` — writes one default `AccountCredentials` blob
  (was 13 deletes + 7 set_news + 1 sync = 21 IPCs; now 2 + 1).
- `Account::read` — calls `read_or_migrate`; copies fields out of
  the loaded creds. Eliminates the 20-tuple match on per-key
  reads and the `parse().unwrap_or_else(...)` defensive ladder
  added in #50 (now centralized in the migration mapping).
- `Account::link` — mutates in-memory state, snapshots the full
  `AccountCredentials`, calls `persist_credentials` once
  (was 18 set_news + 1 sync = 37 IPCs; now 2 + 1). Three
  write-only fields (`account_entropy_pool`, `registration_id`,
  `pni_registration_id`) reach disk via the snapshot without
  needing to be in-memory `Account` fields.
- `Account::set_number` — mutates in-memory, snapshots, persists.
  Cost is one full blob rewrite per call; `set_number` is rare
  (dead-code-flagged path).
- Removes the per-key constants, `get`/`set`/`set_new`
  freestanding helpers, and the 20-tuple `read` matcher. ~280
  lines deleted.

Schema-version compatibility note: registration_id/
pni_registration_id are stored as `u16` (not `u32`) to match the
generator's actual return type from `account_attrs`.

Test count: 175 → 210 visible green; 13 ignored stay ignored.

Precursor target verified: `cargo +stable build --release
--target=riscv32imac-unknown-xous-elf --bin xous-signal-client
--features precursor` clean. Size cost from `bytecheck`: +21 KiB
(420.3 → 441.1 KiB on `xous_signal_client` text). Per-crate
budget remains green; libcrux_ml_kem already over budget on main
(unrelated). Total LOAD VM 4.2 MiB unchanged (matches main).

Notes: xous-signal-client-notes/_open-followups/active/blob-credentials/

Refs #51

+++ src/account/credentials.rs
@@ -53,8 +53,10 @@ pub struct AccountCredentials {
 
     // -- Numeric / boolean -------------------------------------------
     pub device_id: u32,
-    pub registration_id: Option<u32>,
-    pub pni_registration_id: Option<u32>,
+    /// Signal registration IDs are 14-bit per the protocol; stored as
+    /// `u16` to match `account_attrs::generate_registration_id`.
+    pub registration_id: Option<u16>,
+    pub pni_registration_id: Option<u16>,
     pub is_multi_device: bool,
     pub registered: bool,
     pub store_last_receive_timestamp: i64,
@@ -298,15 +300,15 @@ mod tests {
     fn boundary_numerics_round_trip() {
         let mut creds = AccountCredentials::default();
         creds.device_id = u32::MAX;
-        creds.registration_id = Some(u32::MAX);
-        creds.pni_registration_id = Some(u32::MAX);
+        creds.registration_id = Some(u16::MAX);
+        creds.pni_registration_id = Some(u16::MAX);
         creds.store_last_receive_timestamp = i64::MAX;
         creds.store_manifest_version = i64::MIN;
         let bytes = creds.serialize().expect("serialize");
         let back = AccountCredentials::deserialize(&bytes).expect("deserialize");
         assert_eq!(back.device_id, u32::MAX);
-        assert_eq!(back.registration_id, Some(u32::MAX));
-        assert_eq!(back.pni_registration_id, Some(u32::MAX));
+        assert_eq!(back.registration_id, Some(u16::MAX));
+        assert_eq!(back.pni_registration_id, Some(u16::MAX));
         assert_eq!(back.store_last_receive_timestamp, i64::MAX);
         assert_eq!(back.store_manifest_version, i64::MIN);
     }
